Rename Node app path
