+++ nodejs/.npmignore
